# reduce the number of allocations required by Open